+++ crypto/codec/amino.go
@@ -7,6 +7,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec/legacy"
 
 	"github.com/realiotech/realio-network/crypto/ethsecp256k1"
+	"github.com/realiotech/realio-network/crypto/ossecp256k1"
 )
 
 // RegisterCrypto registers all crypto dependency types with the provided Amino
@@ -16,6 +17,9 @@ func RegisterCrypto(cdc *codec.LegacyAmino) {
 		ethsecp256k1.PubKeyName, nil)
 	cdc.RegisterConcrete(&ethsecp256k1.PrivKey{},
 		ethsecp256k1.PrivKeyName, nil)
+
+	cdc.RegisterConcrete(&ossecp256k1.PubKey{}, ossecp256k1.PubKeyName, nil)
+	cdc.RegisterConcrete(&ossecp256k1.PrivKey{}, ossecp256k1.PrivKeyName, nil)
 
 	// NOTE: update SDK's amino codec to include the ethsecp256k1 keys.
 	// DO NOT REMOVE unless deprecated on the SDK.